+++ riscv-peripheral/src/plic/pendings.rs
@@ -30,7 +30,7 @@ impl PENDINGS {
     /// Checks if an interrupt triggered by a given source is pending.
     #[inline]
     pub fn is_pending<I: ExternalInterruptNumber>(self, source: I) -> bool {
-        let source = source.number() as usize;
+        let source = source.number();
         let offset = (source / u32::BITS as usize) as _;
         // SAFETY: valid interrupt number
         let reg: Reg<u32, RO> = unsafe { Reg::new(self.ptr.offset(offset)) };
